change rate column type

+++ database/migrations/2024_02_26_025415_create_transactions_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->decimal('quantity');
             $table->decimal('fee');
-            $table->decimal('rate');
+            $table->float('rate', 10, 4);
             $table->decimal('bought');
             $table->foreignId('from')->constrained('currencies');
             $table->foreignId('to')->constrained('currencies');
